+++ packages/@expo/metro-config/CHANGELOG.md
@@ -40,6 +40,7 @@
 - Point Metro's `assetRegistryPath` at `react-native/asset-registry`, which replaces the `@react-native/assets-registry` package as of React Native 0.87. ([#47729](https://github.com/expo/expo/pull/47729) by [@gabrieldonadel](https://github.com/gabrieldonadel))
 - Bump to `@expo/metro@56.1.0` and `metro@0.84.6` ([#49671](https://github.com/expo/expo/pull/49671) by [@robhogan](https://github.com/robhogan))
 - Don't rename require in tests/snapshots. ([#50039](https://github.com/expo/expo/pull/50039) by [@robhogan](https://github.com/robhogan))
+- Don't rename require in transform worker tests. ([#50143](https://github.com/expo/expo/pull/50143) by [@robhogan](https://github.com/robhogan))
 
 ## 57.0.7 - 2026-07-22